helpers: RetroArch core options improvement
Modified `setRetroarchCoreOption` so it can work when the global options are turned off (default in RetroArch) by either using a 3rd parameter for the core name or by determining the core name from the current module context.

This works by using the module name (`$md_id`) to get the libretro core file then finding the core info file that for that core in order to read the (friendly) core name. The core name is then used to build the `.opt` options file path where the options are saved.