raspbiantools: replace Wayfire with Labwc
RaspiOS now recommends Labwc as a Wayland compositor and it's the default in new installations, replacing Wayfire.
Seems the new compositor is better behaved on memory constrained systems, so it might be suited also for Pi3/4 with 1Gb of RAM, but let's wait until it gets wider testing before removing the Pi4/5 gating.