sdl2: improve checking for updates
Since we didn't have any repo/branch defined, installing the package from source would lead to a 'cannot check for updates' message for this package. Added the repo and branch to to improve this, since we don't update our branch so often and would needlesly re-build and install SDL2 whenever an update was triggered.

Also, don't assume binaries are only for 32bit Arm, we might build binaries for other architectures. Use the native system's arch when looking up for binaries.