xpad: fix build for older kernels
The recent rebasing of the `xpad` driver onto the latest upstream version has added a new feature present only on 6.5+ kernels.
Back out that change so that we can still use the driver on older kernels. This shold be a temporary solution until upstream fixes the compatibility, which may take some time.